requirements.txt
aioquic==1.2.0
annotated-types==0.7.0
anyio
asgiref
attrs==25.1.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
exceptiongroup==1.2.2
fastapi
Flask==3.1.0
h11==0.14.0
h2==4.1.0
hpack==4.1.0
html2text==2024.2.26
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kaitaistruct==0.10
ldap3==2.9.1
lxml==5.3.1
MarkupSafe==3.0.2
mitmproxy-macos==0.10.7
mitmproxy_rs==0.10.7
msgpack==1.1.0
outcome==1.3.0.post0
packaging==24.2
passlib==1.7.4
publicsuffix2==2.20191221
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser
pydantic
pydantic_core
pylsqpack==0.3.19
pyOpenSSL==24.3.0
pyparsing==3.2.0
pyperclip==1.9.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
selenium==4.17.2
service-identity==24.2.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.36.3
tornado==6.4.2
trio==0.29.0
trio-websocket==0.11.1
typing_extensions==4.11.0
undetected-chromedriver==3.5.5
urllib3==2.3.0
urwid==2.6.16
uvicorn==0.27.1
wcwidth==0.2.13
webdriver-manager==4.0.1
websockets==14.2
Werkzeug==3.1.3
wsproto==1.2.0
zstandard==0.23.0
